Section 9 (91 - 96) : Ability System Blueprint Library, AM Widget Controller, Info Delegate, Widget Attribute Tags
Created the AuraAbilitySystemLibrary, Constructed the Attribute Widget Controller, Added Attribute Info Delegates and broadcasted said info. Created Attribute Tags for each of the widgets that correspond to each attribute. Created a map and mapped the tags to the get attribute functions. Made it so there is a response to when an attribute is changed.