Section 15 (166 - 168) Added Blackboard keys and Player/Enemy Tag, Worked on BTService Find Nearest Player class.
Made sure enemies rotated using the charactermovement component rather than the controller rotation. Added bools in the BTService to allow for enemies to either be or not be affected by gameplay effects such as being able to be damaged by ground fire or not picking up items that the player can only do. Fixed minor bug with ghost bar not fully catching up with the health. Fixed Mana and health potions/crystals.