Section 16 ( 177 - 179) Attack Gameplay Ability, Melee Attack Montage and Combat Target for the Enemy
New Attack Tag was created to allow enemies to know when to trigger an attack. Added the ability for the enemies to have startup abilities where it checks which type of enemy it is and gives the gameplay ability; in this case the MeleeAttack ability was added to the Warrior class. Created a Montage and a Warp Motion where it sets or gets the Combat Target for melee enemies to face in the direction of said targer when attacking.