Section 19 (211 - 219) Shaman/Demon/Ghoul Sounds and Notifies. New Ability, Summons.
Created new enemy (Demon) that functions as a summon minion for the shaman. Demon has two classes, a range and melee class. Added Sounds and Notifies for Shaman, Demon and Ghoul. Added Swipe trail for Ghoul when attacking. Created new skeleton Gameplay Ability to summon minions in an angle range in front of the Shaman.