Section 19 (220 - 228) Added: Async Spawn Time, Particle Effects, Elementalist Behavior Tree, Elementalist Attack Task, Minion Tweening
Added Async Spawn Time for summoning so it is semi-random and not consistent on the location around the fan area of the goblin summoner. Added Particle Effects for the summoning and Minion Scale Tweening so it gives it juice. Added new Behavior Tree for the Elementalist goblin so it can shoot and summon minions depending on how many it has summoned. Added New Attack Task for Elementalist that allows for the previous statement to happen.