Section 25 (302 - 314): Debuffs, Dynamic Gameplay Effects, Death Impulse, Knockback
Added scalable code for debuffs and created a new type of debuff (burn) that will be applied on fire damage depending on the parameters. Added also death impulse which will impulse enemies and character on death and added the ability to knockback. Fixed bugs that did not allow hit reacting during the debuff tag which was turned into a boolean for gameplay purposes. Also fixed a bug on knockback which did not function correctly due to the character moving while being knockedback which resulted in the character getting sent into the air rather than pushed back.