Secttion 26 (318 - 327) Aura Beam Spell, Electroute Looping Cue, First Trace Target
Added Niagara System for when clicking on the floor. Added new beam spell with Niagara system and Montage. Added block tags to disable cursor trace, input held, and input pressed while electrocute is casting. Added Gameplay Cue for electrocute looping sound. Made New Gameplay Cue for electrocute remove and add the cue to specific actors, if it hits the floor for example, the cue will be added and removed from aura, but if it hits another actor for example an enemy, the cue will be added to that enemy and removed from said enemy as this spell will chain to other enemies in the future.